Add context awareness
By putting `case <-ctx.Done():` in a branch of a select block, we can:
1. Protect from deadlocks, for instance when writing to an unbuffered channel with no reader.
2. Gracefully end our goroutines. For instance, finish processing the last message before exiting.

If we want channel reads to be interruptible, then we can't use nice syntax things like `for msg := range myChannel`, because context is just a library, not something built in to the language. But I think it's pretty much always worth it, because it's too easy to get things wrong and cause deadlocks.