Introduce a "router" style PubSub
I think this is a better option than the "Runner" style. More of the channel and goroutine specifics are hidden in the Router implementation, rather than exposed as part of the API to the consumer. In the "runner" case, the consumer needs to deal with the publish channel it was provided, it needs to create a channel for each subscribe, a goroutine for each subscriber, etc. IMO this saddles the consumer with a lot of responsibility for thinking about how the pubsub works. The only place where the router implementation exposes the existence of channels or goroutines is that the listener passed to the subscriber gets new messages through a provided channel. The moral of the story is that go concurrency patterns work best when the concurrency primitives are used primarily as implementation details, not exposed as part of the API or contract.

A heuristic that follows from this moral: channels should be provisioned near wherever corresponding goroutines are spawned.